requirements.txt
fastapi
pydantic
appwrite
python-dotenv
uvicorn
python-multipart